fixed todolist
