+++ src/components/TodosFilter/TodosFilter.tsx
@@ -1,12 +1,10 @@
 import { useContext } from 'react';
-import { DispatchContext, Status, Todo, TodosContext } from '../../store';
+import { DispatchContext, Status, TodosContext } from '../../store';
 import classNames from 'classnames';
 
 export const TodosFilter = () => {
-  const { todos, status } = useContext(TodosContext);
+  const { status } = useContext(TodosContext);
   const dispatch = useContext(DispatchContext);
-  const completed = todos.filter((todo: Todo) => todo.completed);
-  const active = todos.filter((todo: Todo) => !todo.completed);
 
   const handleStatusChange = (newStatus: Status) => {
     dispatch({
@@ -15,64 +13,39 @@ export const TodosFilter = () => {
     });
   };
 
-  const handleClearAllCompleted = () => {
-    completed.forEach((todo: Todo) => {
-      dispatch({
-        type: 'remove',
-        id: todo.id,
-      });
-    });
-  };
-
   return (
-    <footer className="footer" data-cy="todosFilter">
-      <span className="todo-count" data-cy="todosCounter">
-        {active.length} items left
-      </span>
-
-      <ul className="filters">
-        <li>
-          <a
-            href={`#/${Status.All}`}
-            className={classNames('', { selected: status === Status.All })}
-            onClick={() => handleStatusChange(Status.All)}
-          >
-            All
-          </a>
-        </li>
-
-        <li>
-          <a
-            href={`#/${Status.Active}`}
-            className={classNames('', { selected: status === Status.Active })}
-            onClick={() => handleStatusChange(Status.Active)}
-          >
-            Active
-          </a>
-        </li>
-
-        <li>
-          <a
-            href={`#/${Status.Completed}`}
-            className={classNames('', {
-              selected: status === Status.Completed,
-            })}
-            onClick={() => handleStatusChange(Status.Completed)}
-          >
-            Completed
-          </a>
-        </li>
-      </ul>
-
-      {completed.length > 0 && (
-        <button
-          type="button"
-          className="clear-completed"
-          onClick={handleClearAllCompleted}
+    <ul className="filters" data-cy="todosFilter">
+      <li>
+        <a
+          href={`#/${Status.All}`}
+          className={classNames('', { selected: status === Status.All })}
+          onClick={() => handleStatusChange(Status.All)}
         >
-          Clear completed
-        </button>
-      )}
-    </footer>
+          All
+        </a>
+      </li>
+
+      <li>
+        <a
+          href={`#/${Status.Active}`}
+          className={classNames('', { selected: status === Status.Active })}
+          onClick={() => handleStatusChange(Status.Active)}
+        >
+          Active
+        </a>
+      </li>
+
+      <li>
+        <a
+          href={`#/${Status.Completed}`}
+          className={classNames('', {
+            selected: status === Status.Completed,
+          })}
+          onClick={() => handleStatusChange(Status.Completed)}
+        >
+          Completed
+        </a>
+      </li>
+    </ul>
   );
 };
